Status components in a hook
